requirements.txt
Django==4.2.17
twilio==9.3.7
djangorestframework==3.15.2
python-decouple==3.8
gunicorn==23.0.0
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
psycopg2-binary==2.9.10
google-api-python-client==2.151.0
google-auth-oauthlib==1.2.1
pytz==2024.2
